+++ client/src/App.tsx
@@ -4,12 +4,13 @@ import { Routes, Route } from 'react-router-dom';
 import LoginPage from './pages/LoginPage/LoginPage';
 import HomePage from './pages/HomePage/HomePage';
 import RegisterPage from './pages/RegisterPage/RegisterPage';
+import LoadingPage from './pages/LoadingPage/LoadingPage';
 
 const App = () => {
 	return (
 		<Routes>
 			<Route path="/" element={<LoginPage />} />
-			<Route path="/oauth/naverCallback" element={<LoginPage />} />
+			<Route path="/oauth/naverCallback" element={<LoadingPage />} />
 			<Route path="/home" element={<HomePage />} />
 			<Route path="/register" element={<RegisterPage />} />
 		</Routes>
