+++ client/src/pages/LoginPage/LoginPage.tsx
@@ -1,5 +1,4 @@
-import React, { useEffect } from 'react';
-import axios from 'axios';
+import React from 'react';
 // style
 import { LoginPageContainer, OauthButton } from './LoginPageStyles';
 // img
@@ -16,46 +15,6 @@ const LoginPage = () => {
 		const naverLoginUrl = `https://nid.naver.com/oauth2.0/authorize?response_type=code&client_id=${clientId}&state=${stateString}&redirect_uri=${callbackUrl}`;
 		window.location.assign(naverLoginUrl);
 	};
-
-	const getSocialName = (url: URL) => {
-		const callback = url.pathname.split('/')[2];
-		let name = '';
-		if (callback.includes('naver')) {
-			name = 'naver';
-		} else if (callback.includes('kakao')) {
-			name = 'kakao';
-		}
-		return name;
-	};
-
-	// TODO: authorization code를 BE에 전송 후, reponse에 따른 처리 필요
-	const postAuthrizationInfo = async (
-		socialName: string,
-		authorizationCode: string,
-		state: string,
-	) => {
-		const response = await axios.post(`/api/oauth/${socialName}`, {
-			authorizationCode,
-			state,
-		});
-		if (response.data.code === 200) {
-			if (response.data.email !== undefined) {
-				// TODO: 회원가입
-			} else {
-				// TODO: 로그인되어 홈으로 이동
-			}
-		}
-	};
-
-	useEffect(() => {
-		const url = new URL(window.location.href);
-		const authorizationCode = url.searchParams.get('code');
-		const state = url.searchParams.get('state');
-		if (authorizationCode && state) {
-			const socialName = getSocialName(url);
-			postAuthrizationInfo(socialName, authorizationCode, state);
-		}
-	}, []);
 
 	return (
 		<LoginPageContainer>
